refactor: refactoring new post page
