+++ src/utils/getRandomColors.ts
@@ -1,0 +1,4 @@
+export const getRandomHexColor = () => {
+    const randomColor = Math.floor(Math.random() * 16777215).toString(16);
+    return `#${randomColor.padStart(6, '0')}`;
+  };
